@@ -18,19 +18,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-#  syncfusion_flutter_calendar:
-#    path: E:\jan 20\flutter-calendar\flutter_calendar\syncfusion_flutter_calendar
-  syncfusion_flutter_datepicker: 18.4.43-beta
-  intl: ^0.16.1
+
+  syncfusion_flutter_datepicker: ^19.1.55-beta
+  intl: ^0.17.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.1
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
